web/requirements.txt
dash
dash_bootstrap_components
gunicorn
azure-servicebus
azure-data-tables
azure-core